Fix JDK macro
Since Java 17, the links to the JDK classes must include their module.
This commit fixes the macro so that all classes which package starts
with `java` are assigned the `java.base` module automatically, but
makes it possible to specify their module explicitly by using the
`[module=....]` attribute.

Fixes #605
