+++ src/main/groovy/io/micronaut/docs/Jdk.groovy
@@ -4,7 +4,7 @@ import groovy.transform.CompileStatic
 
 @CompileStatic
 class Jdk implements JvmLibrary {
-    private static final String DEFAULT_URI = "https://docs.oracle.com/javase/17/docs/api"
+    private static final String DEFAULT_URI = "https://docs.oracle.com/en/java/javase/17/docs/api"
 
     @Override
     String defaultUri() {
