+++ src/test/groovy/io/micronaut/docs/ConfigurationPropertiesMacroSpec.groovy
@@ -1,33 +1,6 @@
 package io.micronaut.docs
 
-import org.asciidoctor.Asciidoctor
-import org.asciidoctor.Attributes
-import org.asciidoctor.Options
-import org.asciidoctor.SafeMode
-import spock.lang.Specification
-
-class ConfigurationPropertiesMacroSpec extends Specification {
-    private Asciidoctor asciidoctor
-    private String converted
-    private Options options
-
-    def setup() {
-        asciidoctor = Asciidoctor.Factory.create()
-        asciidoctor.javaExtensionRegistry().block(new ConfigurationPropertiesMacro(asciidoctor))
-        options = Options.builder()
-                .safe(SafeMode.SAFE)
-                .attributes(
-                        Attributes.builder()
-                                .attribute('source-highlighter', 'highlightjs')
-                                .build()
-                )
-                .backend("html5")
-                .build()
-    }
-
-    def cleanup() {
-        asciidoctor.shutdown()
-    }
+class ConfigurationPropertiesMacroSpec extends AbstractConverterSpec {
 
     def "converts properties"() {
         when:
@@ -116,7 +89,4 @@ mongodb {
 </div>'''
     }
 
-    void convert(String input) {
-        converted = asciidoctor.convert(input, options)
-    }
 }
